Changed project structure to use `src/` directory as the main directory code for better understanding and improve project structure
