--- app/(tabs)/stats/index.tsx
@@ -1,18 +1,0 @@
-import {StyleSheet, View} from "react-native";
-import {ThemedText} from "@/components/ThemedText";
-
-export default function Page() {
-  return (
-      <View style={styles.container}>
-        <ThemedText>Stats Screen</ThemedText>
-      </View>
-  )
-}
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        justifyContent: "center",
-        alignItems: "center",
-    }
-});
