+++ src/assets/images/adaptive-icon.png
