+++ src/components/screens/home/meds-cards.tsx
@@ -1,10 +1,10 @@
-import {FlatList, Platform, Text, View} from "react-native";
+import {FlatList, Text, View} from "react-native";
 import {ThemedText} from "@/components/ThemedText";
 import {Image} from "expo-image";
 import {Checkbox} from "expo-checkbox";
 import {useState} from "react";
 import {cssInterop} from "nativewind";
-import {cn} from "@/components/tailwind-merge";
+import {cn} from "@/utils/tailwind-merge";
 
 interface SingleMedCardProps {
     id?: string;
@@ -24,8 +24,8 @@ const StyledImage = cssInterop(Image, {
     className: 'style',
 })
 
-export function MedCard({imageURI, med, dosage, nextDose}: SingleMedCardProps) {
-    const defaultImageURI = imageURI || 'https://static.vecteezy.com/system/resources/previews/005/259/960/non_2x/medicine-drugs-box-free-vector.jpg';
+export function MedCard({ card }: { card: SingleMedCardProps } ) {
+    const defaultImageURI = card.imageURI || 'https://static.vecteezy.com/system/resources/previews/005/259/960/non_2x/medicine-drugs-box-free-vector.jpg';
     const [checked, setChecked] = useState(false);
 
     return (
@@ -34,14 +34,14 @@ export function MedCard({imageURI, med, dosage, nextDose}: SingleMedCardProps) {
             <View className={'flex-row items-center gap-2'}>
                 <StyledImage source={{uri: defaultImageURI}} className={'w-[80] h-[80] rounded-xl'}/>
                 <View className={'flex-col py-2 justify-between'}>
-                    <ThemedText type={'subtitle'}>{med}</ThemedText>
-                    <Text className={'text-gray-600 text-xl'}>{dosage}</Text>
+                    <ThemedText type={'subtitle'}>{card.med}</ThemedText>
+                    <Text className={'text-gray-600 text-xl'}>{card.dosage}</Text>
                 </View>
             </View>
 
             <View className={'flex-row justify-center items-center gap-2 w-1/3'}>
                 <Checkbox value={checked} onValueChange={setChecked}/>
-                <ThemedText>{nextDose}</ThemedText>
+                <ThemedText>{card.nextDose}</ThemedText>
             </View>
         </View>
     );
@@ -58,12 +58,7 @@ export function MedCards({MedCardsArray, className}: MedCardsProps) {
             )}
             data={MedCardsArray}
             renderItem={({item}) => (
-                <MedCard
-                    imageURI={item.imageURI}
-                    med={item.med}
-                    dosage={item.dosage}
-                    nextDose={item.nextDose}
-                />
+                <MedCard card={item} />
             )}
             keyExtractor={(item) => {
                 if (item.id === undefined) {
