+++ src/components/ui/TabBarBackground.ios.tsx
